feat: 상세페이지 og태그 메타태그 추가
